requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-core==2.41.5
SQLAlchemy==2.0.28
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-multipart==0.0.22
fpdf2==2.8.5
requests>=2.28
zeep>=4.1.0

# Password hashing (required by app/main.py)
passlib[bcrypt]>=1.7.4
# bcrypt backend (optional explicit pin)
bcrypt>=4.0.0
# JWT handling
python-jose[cryptography]>=3.3.0
